+++ programs/neovim/config/lua/user/files/init.lua
@@ -1,3 +1,4 @@
 require("user.files.c")
 require("user.files.ecmascript")
 require("user.files.groovy")
+require("user.files.netrw")