+++ programs/neovim/config/lua/user/files/netrw.lua
@@ -1,0 +1,8 @@
+vim.api.nvim_create_autocmd({
+	"FileType",
+}, {
+	pattern = "netrw",
+	callback = function()
+		vim.opt.relativenumber = true
+	end,
+})